Move upload code to a function
